requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-genai==1.14.0
requests==2.32.3
beautifulsoup4==4.13.3
